# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/topup": {
            "post": {
                "tags": [
                    "topup"
                ],
                "summary": "Create a new topup transaction",
                "description": "Create a new user topup transaction with user_id, wallet_id, coin_amount and price",
                "operationId": "topup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "wallet_id",
                                    "coin_amount",
                                    "price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "user_uuid"
                                    },
                                    "wallet_id": {
                                        "type": "string",
                                        "example": "wallet_uuid"
                                    },
                                    "coin_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Topup transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Topup transaction created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "topup_uuid"
                                                },
                                                "wallet_id": {
                                                    "type": "string",
                                                    "example": "wallet_uuid"
                                                },
                                                "coin_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100000
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Proses"
                                                },
                                                "created_by": {
                                                    "type": "string",
                                                    "example": "userA"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-07-07 15:50:43"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/topup-list": {
            "get": {
                "tags": [
                    "topup"
                ],
                "summary": "Get list of all topup transactions",
                "description": "Retrieve all user topup transactions ordered by updated_at descending",
                "operationId": "listTopupTransactions",
                "responses": {
                    "200": {
                        "description": "List of topup transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of topup transactions retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "topup_uuid"
                                                    },
                                                    "wallet_id": {
                                                        "type": "string",
                                                        "example": "wallet_uuid"
                                                    },
                                                    "coin_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 100000
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15000
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Proses"
                                                    },
                                                    "created_by": {
                                                        "type": "string",
                                                        "example": "userA"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-07-07 15:50:43"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-07-07 15:55:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/topup/{id}/cancel": {
            "post": {
                "tags": [
                    "topup"
                ],
                "summary": "Cancel a topup transaction",
                "description": "Cancel a user topup transaction by ID",
                "operationId": "cancelTopupTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the topup transaction to cancel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topup transaction cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Topup transaction cancelled successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "topup_uuid"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Cancelled"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Topup transaction not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Topup transaction not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}